Implement view component for power service. Fix brightnessservice in dal ambientlight unit.

+++ remote/src/main/java/de/citec/dal/remote/unit/BrightnessSensorRemote.java
@@ -30,6 +30,6 @@ public class BrightnessSensorRemote extends DALRemoteService<BrightnessSensorTyp
 
     @Override
     public Double getBrightness() throws CouldNotPerformException {
-        return (double) this.getData().getBrightness();
+        return this.getData().getBrightness();
     }
 }
